requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
isodate==0.6.1
pydantic==1.9.0
pyparsing==3.0.7
PyYAML==6.0
rdflib==6.1.1
requests==2.27.1
six==1.16.0
typing-extensions==4.0.1
urllib3==1.26.8
